cmake: auto-detect LIB_SUFFIX, allowing for manual override
On x86_64 the lib will be installed in lib64 dir. We use GNUInstallDirs
cmake-module for detecting that. In case it is not available, a warning
with instructions will be given to manually set the suffix.
Either way, if LIB_SUFFIX is manually given (for example by system
packaging tools), it will override the auto-detection.