driver: modify JacoArm constructor/destrucor to handle arbritary amount of arms
Upon each constructor call, all the currently connected arms are checked.
Those information might be already known, in case an arm has not been
removed since. Those known devices also have a flag that tells if they are
in use, such that from all connected arms it is possible to find one
that is not in use (if actually available).

Upon destruction, the information is stored that this arm is available
again.

This allows to call the JacoArm constructor arbitrary times, each time
checking for a new "free" arm.