driver: flush packets on incoming port after connecting to arm
There might be still some packets left on the device that it will send
with the next request, probably due to some crash in your software while
communicating with the arm. If we don't flush them, they will be received
first when the next command is sent, which will probably lead to a
CMD_ID_MISMATCH error.