driver: Add support for Jaco tactile sensors
The new class JacoSensors can be used to connec to the Jaco
tactile sensors and fetch data.
Raw and normalized pressure data for each sensor, as well as
the vibration on the sensor surface are the currently supported
methods. Methods to get multiple sensor data at once will be added
in the short future.

The KinDrv library will automatically build the support for the
sensors. The included udev-rules have been extended to install
a symlink for the connected sensor.